# target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":""},"host":"localhost:8080","basePath":"/NfvDeployer/rest","tags":[{"name":"catalog"},{"name":"nffgs"},{"name":"host"},{"name":"connections"}],"schemes":["http"],"paths":{"/catalog":{"get":{"tags":["catalog"],"summary":"get the catalog","description":"return to the client the entire catalog of VNF that are available into the web service","operationId":"getCatalog","produces":["application/xml"],"parameters":[],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"},"204":{"description":"No Content"}}}},"/connections/{host1Id}/{host2Id}":{"get":{"tags":["connections"],"summary":"get connnection performance","description":"return the connection performace parameters between two host into the system","operationId":"getConnections","produces":["application/xml"],"parameters":[{"name":"host1Id","in":"path","required":true,"type":"string"},{"name":"host2Id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"}}}},"/hosts":{"get":{"tags":["host"],"summary":"get the hosts collection","description":"return all host that are available inside the web service","operationId":"getHosts","produces":["application/xml"],"parameters":[],"responses":{"201":{"description":"OK"},"500":{"description":"Internal Server Error"},"204":{"description":"No Content"}}}},"/hosts/{hostId}":{"get":{"tags":["host"],"summary":"get a single host","description":"return to the client the representation of a single host","operationId":"getHost","produces":["application/xml"],"parameters":[{"name":"{hostId}","in":"path","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"}}}},"/nffgs":{"get":{"tags":["nffgs"],"summary":"get nffgs collections","description":"return the xml representation of the nffg graphs deployed into the system","operationId":"getNffgs","produces":["application/xml"],"parameters":[{"name":"date","in":"query","required":false,"type":"string","format":"date-time"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"},"204":{"description":"No Content"}}},"post":{"tags":["nffgs"],"summary":"create a new nffg","description":"deploy a new nffg into the infrastructure network","operationId":"postNewNffg","consumes":["application/xml"],"produces":["text/plain"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/JAXBElementNffgGraphType"}}],"responses":{"201":{"description":"Created"},"400":{"description":"Bad Request"},"500":{"description":"Internal Server Error"},"404":{"description":"Forbidden"}}}},"/nffgs/{nffgId}":{"get":{"tags":["nffgs"],"summary":"get a single nffg","description":"return the xml representation of the nffg graphs deployed into the system","operationId":"getNffg","produces":["application/xml"],"parameters":[{"name":"nffgId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"}}},"delete":{"tags":["nffgs"],"summary":"delete a nffg","description":"delete an nffg and his graph","operationId":"deleteNffgs","parameters":[],"responses":{"500":{"description":"Internal Server Error"},"501":{"description":"Not Implemented"},"404":{"description":"Not Found"}}}},"/nffgs/{nffgId}/links":{"post":{"tags":["nffgs"],"summary":"create a new link","description":"create a new link inside the graph","operationId":"createNewLink","consumes":["application/xml"],"produces":["text/plain"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/JAXBElementExtendedLinkType"}},{"name":"nffgId","in":"path","required":true,"type":"string"}],"responses":{"201":{"description":"Created"},"500":{"description":"Internal Server Error"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/nffgs/{nffgId}/links/{linkId}":{"delete":{"tags":["nffgs"],"summary":"delete a link","description":"delete a nf-fg link","operationId":"deleteLink","parameters":[{"name":"linkId","in":"path","required":true,"type":"string"},{"name":"nffgId","in":"path","required":true,"type":"string"},{"name":"nodeId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Not Implemented"},"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"}}}},"/nffgs/{nffgId}/nodes":{"post":{"tags":["nffgs"],"summary":"create a node","description":"delete a nf-fg node","operationId":"createNewNode","consumes":["application/xml"],"produces":["text/plain"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/JAXBElementNodeType"}},{"name":"nffgId","in":"path","required":true,"type":"string"}],"responses":{"201":{"description":"Created"},"500":{"description":"Internal Server Error"},"404":{"description":"Forbidden"}}}},"/nffgs/{nffgId}/nodes/{nodeId}":{"delete":{"tags":["nffgs"],"summary":"delete a node","description":"delete a nf-fg node","operationId":"deleteNode","parameters":[{"name":"nodeId","in":"path","required":true,"type":"string"},{"name":"nffgId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Not Implemented"},"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"}}}},"/nffgs/{nffgId}/nodes/{nodeId}/reachableHost":{"get":{"tags":["nffgs"],"summary":"get all the reachable hosts ","description":"get the hosts that are reachable from the specified node","operationId":"getReachableNodes","produces":["application/xml"],"parameters":[{"name":"nffgId","in":"path","required":true,"type":"string"},{"name":"nodeId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"},"204":{"description":"No Body"},"404":{"description":"Not Found"}}}}},"definitions":{"JAXBElementNffgGraphType":{"type":"object","properties":{"name":{"$ref":"#/definitions/QName"},"value":{"$ref":"#/definitions/NffgGraphType"},"nil":{"type":"boolean","default":false},"globalScope":{"type":"boolean","default":false},"typeSubstituted":{"type":"boolean","default":false}}},"NffgGraphType":{"type":"object","required":["deployDate","links","nffgName","nodes"],"properties":{"nffgName":{"type":"string","xml":{"name":"nffg-name"}},"deployDate":{"type":"string","format":"date-time","xml":{"name":"deploy-date"}},"nodes":{"$ref":"#/definitions/Nodes"},"links":{"$ref":"#/definitions/Links"}}},"JAXBElementExtendedLinkType":{"type":"object","properties":{"name":{"$ref":"#/definitions/QName"},"value":{"$ref":"#/definitions/ExtendedLinkType"},"nil":{"type":"boolean","default":false},"globalScope":{"type":"boolean","default":false},"typeSubstituted":{"type":"boolean","default":false}}},"QName":{"type":"object","properties":{"namespaceURI":{"type":"string"},"localPart":{"type":"string"},"prefix":{"type":"string"}}},"NodeType":{"type":"object","required":["hostname","name","nfFg","vnf"],"properties":{"link":{"type":"array","items":{"$ref":"#/definitions/LinkType"}},"name":{"type":"string"},"vnf":{"type":"string","xml":{"name":"VNF"}},"hostname":{"type":"string"},"nfFg":{"type":"string","xml":{"name":"nf-fg"}}}},"Links":{"type":"object","required":["link"],"properties":{"link":{"type":"array","items":{"$ref":"#/definitions/ExtendedLinkType"}}}},"LinkType":{"type":"object","required":["destinationNode","linkName","sourceNode"],"properties":{"linkName":{"type":"string","xml":{"name":"link-name"}},"destinationNode":{"type":"string"},"sourceNode":{"type":"string"},"latency":{"type":"integer","format":"int64"},"throughput":{"type":"number","format":"float"}}},"JAXBElementNodeType":{"type":"object","properties":{"name":{"$ref":"#/definitions/QName"},"value":{"$ref":"#/definitions/NodeType"},"nil":{"type":"boolean","default":false},"globalScope":{"type":"boolean","default":false},"typeSubstituted":{"type":"boolean","default":false}}},"ExtendedLinkType":{"type":"object","required":["destinationNode","linkName","sourceNode"],"properties":{"linkName":{"type":"string","xml":{"name":"link-name"}},"destinationNode":{"type":"string"},"sourceNode":{"type":"string"},"latency":{"type":"integer","format":"int64"},"throughput":{"type":"number","format":"float"},"overwrite":{"type":"boolean","default":false}}},"Nodes":{"type":"object","required":["node"],"properties":{"node":{"type":"array","items":{"$ref":"#/definitions/NodeType"}}}}}}